LPS-200689 partially move UserGroupRoleModelListener with _clearCache() to modules and make it a component
